requirements.txt
astroid==2.3.3
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
Django==2.2
django-autofixture==0.12.1
django-ckeditor==5.9.0
django-embed-video==1.3.1
django-js-asset==1.2.2
django-suit==0.2.28
django-user-agents==0.4.0
gunicorn==20.0.4
idna==2.9
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
Pillow==6.0.0
py3dns==3.2.1
pylint==2.4.4
pytz==2019.1
PyYAML==5.3
requests==2.23.0
six==1.14.0
sqlparse==0.3.0
typed-ast==1.4.1
ua-parser==0.10.0
urllib3==1.25.8
user-agents==2.1
validate-email==1.3
whitenoise==4.1.2
wrapt==1.11.2
